requirements.txt
flask
faiss-cpu
sentence-transformers
neo4j
langchain
langgraph
